Moves Parser into commons, thus making it available via the API
